+++ mail/mail-uniffi/CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [ET-2368] Saving registered device is now generating new device token key pair if necessary.
   - [ET-2368] Registering device is now using device key instead of user key.
 
+### Added
+
+  - [ET-2394] Expose message banners (new field in `WatchedConversation`)
+
 ## [0.55.0] - 2025-03-07
 
 ### Added